KG_XGBoost/requirements.txt
aiobotocore==2.4.2
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
alabaster==0.7.12
alembic==1.10.2
anaconda-client==1.7.2
anaconda-project==0.8.3
ansi2html==1.8.0
anyio==3.6.2
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.3.0
astroid==2.15.0
astropy==4.3.1
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.3.0
attrs==22.2.0
autopep8==1.4.4
autovizwidget @ file:///home/conda/feedstock_root/build_artifacts/autovizwidget_1675144657283/work
awscli==1.27.96
Babel==2.12.1
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
bleach==6.0.0
bokeh==1.4.0
boto==2.49.0
boto3==1.26.96
botocore==1.29.96
Bottleneck==1.3.2
brotlipy==0.7.0
cached-property==1.5.2
certifi==2022.12.7
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chardet==3.0.4
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
Click==7.0
cloudpickle==2.2.1
clyent==1.2.2
cmaes==0.9.1
colorama==0.4.3
colorlog==6.7.0
conda==22.9.0
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1649105789509/work
contextlib2==0.6.0.post1
cryptography==39.0.2
cycler==0.10.0
Cython==0.29.15
cytoolz==0.10.1
dash==2.9.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2022.2.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.6
distributed==2022.2.0
distro==1.8.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
dparse==0.6.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
fastjsonschema==2.16.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.2
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
google-pasta==0.2.0
greenlet==0.4.15
h5py==2.10.0
hdijupyterutils @ file:///home/conda/feedstock_root/build_artifacts/hdijupyterutils_1675108186120/work
HeapDict==1.0.1
html5lib==1.0.1
huggingface-hub==0.13.3
hypothesis==5.5.4
idna==2.8
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.34.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.2
jeepney==0.4.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.1.0
jupyter-dash==0.4.2
jupyter-server==1.23.6
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab==1.2.21
jupyterlab-pygments==0.2.2
jupyterlab-server==1.0.6
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
lief==0.9.0
llvmlite==0.39.1
locket==0.2.0
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.1.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==6.0.4
multipledispatch==0.6.0
multiprocess==0.70.14
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==6.5.4
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.4
nltk==3.8.1
nose==1.3.7
notebook==6.5.3
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.7.1
numpy==1.21.6
numpydoc==0.9.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
olefile==0.46
openpyxl==3.0.3
optuna==3.1.0
packaging==23.0
pandas==1.3.5
pandocfilters==1.4.2
parso==0.8.3
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathos==0.3.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkginfo==1.5.0.1
platformdirs==3.1.1
plotly==5.8.2
pluggy==0.13.1
ply==3.11
pox==0.3.2
ppft==1.7.6.6
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.20.3
protobuf3-to-dict==0.1.5
psutil==5.6.7
ptyprocess==0.6.0
pure-sasl @ file:///home/conda/feedstock_root/build_artifacts/pure-sasl_1631890804823/work
py==1.11.0
pyarrow==11.0.0
pyasn1==0.4.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle==4.0.1
pyerfa==2.0.0.2
pyflakes==2.1.1
pyfunctional==1.4.3
Pygments==2.14.0
PyHive @ file:///home/conda/feedstock_root/build_artifacts/pyhive_1646707521362/work
pykerberos==1.2.1
pylint==2.17.1
pynndescent==0.5.8
pyodbc===4.0.0-unsupported
pyOpenSSL==23.0.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==6.0
pyzmq==25.0.2
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
regex==2022.10.31
requests==2.28.2
requests-kerberos @ file:///home/conda/feedstock_root/build_artifacts/requests-kerberos_1636815972022/work
retrying==1.3.4
rope==0.16.0
rsa==4.9
Rtree==0.9.3
ruamel_yaml==0.15.87
s3fs==0.4.2
s3transfer==0.6.0
sagemaker==2.140.1
sagemaker-data-insights==0.3.3
sagemaker-datawrangler==0.3.9
sagemaker-scikit-learn-extension==2.5.0
sagemaker-studio-analytics-extension==0.0.17
sagemaker-studio-sparkmagic-lib==0.1.4
sasl==0.3.1
schema==0.7.5
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
SecretStorage==3.1.2
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
smclarify==0.4
smdebug-rulesconfig==1.0.1
sniffio==1.3.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
sparkmagic @ file:///home/conda/feedstock_root/build_artifacts/sparkmagic_1675108208766/work/sparkmagic
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.13
statsmodels==0.11.0
sympy==1.5.1
tables==3.6.1
tabulate==0.9.0
tblib==1.6.0
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1677600641219/work
terminado==0.8.3
testpath==0.4.4
texttable==1.6.7
thrift @ file:///home/conda/feedstock_root/build_artifacts/thrift_1614958308644/work
thrift-sasl @ file:///home/conda/feedstock_root/build_artifacts/thrift_sasl_1664049052220/work
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.10.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.42.1
traitlets==5.9.0
transformers==4.27.3
typed-ast==1.5.4
typing_extensions==4.5.0
ujson==5.7.0
umap-learn==0.5.3
unicodecsv==0.14.1
urllib3==1.26.15
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.2.3
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
xgboost==1.6.2
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
yapf==0.28.0
yarl==1.8.2
zict==1.0.0
zipp==3.15.0
